+++ src/client/resources/assets/vfxweaver/shaders/include/dome.glsl
@@ -11,9 +11,4 @@ vec2 vfx_dome_uv(vec3 direction) {
     return vec2((yaw + 180.0) / 360.0, (pitch + 90.0) / 180.0);
 }
 
-vec2 vfx_dome_uv_rotated(vec3 direction, float rotationDegrees) {
-    float angle = radians(rotationDegrees);
-    float c = cos(angle);
-    float s = sin(angle);
-    return vfx_dome_uv(vec3(c * direction.x - s * direction.z, direction.y, s * direction.x + c * direction.z));
-}
+